Move menu/toolbar logic from MainWindow to separate service.
