+++ ILSpy/MainWindow.xaml.cs
@@ -17,43 +17,36 @@
 // DEALINGS IN THE SOFTWARE.
 
 using System;
-using System.Collections.Generic;
-using System.Collections.Specialized;
 using System.ComponentModel;
 using System.Diagnostics;
+using System.Drawing;
 using System.Globalization;
 using System.IO;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
 using System.Windows.Input;
 using System.Windows.Media;
 using System.Windows.Navigation;
 
 using AvalonDock.Layout.Serialization;
 
-using ICSharpCode.Decompiler;
 using ICSharpCode.ILSpy.AppEnv;
 using ICSharpCode.ILSpy.AssemblyTree;
-using ICSharpCode.ILSpy.Commands;
 using ICSharpCode.ILSpy.Docking;
 using ICSharpCode.ILSpy.Search;
 using ICSharpCode.ILSpy.TextView;
-using ICSharpCode.ILSpy.Themes;
 using ICSharpCode.ILSpy.TreeNodes;
 using ICSharpCode.ILSpy.Updates;
 using ICSharpCode.ILSpy.ViewModels;
 using ICSharpCode.ILSpyX.FileLoaders;
 using ICSharpCode.ILSpyX.Settings;
-using ICSharpCode.ILSpyX.Extensions;
-
-using Microsoft.Win32;
 using ICSharpCode.ILSpyX.TreeView;
 
-using TomsToolbox.Composition;
+using Microsoft.Win32;
+
+using Screen = System.Windows.Forms.Screen;
 
 namespace ICSharpCode.ILSpy
 {
@@ -74,13 +67,6 @@ namespace ICSharpCode.ILSpy
 			get {
 				return App.ExportProvider.GetExportedValue<AssemblyListPaneModel>();
 			}
-		}
-
-		public DecompilationOptions CreateDecompilationOptions()
-		{
-			var decompilerView = DockWorkspace.Instance.ActiveTabPage.Content as IProgress<DecompilationProgress>;
-
-			return new(AssemblyTreeModel.CurrentLanguageVersion, SettingsService.Instance.DecompilerSettings, SettingsService.Instance.DisplaySettings) { Progress = decompilerView };
 		}
 
 		public MainWindow()
@@ -104,31 +90,14 @@ namespace ICSharpCode.ILSpy
 				Thread.CurrentThread.CurrentUICulture = new CultureInfo(sessionSettings.CurrentCulture);
 			}
 			InitializeComponent();
-			InitToolPanes();
+
 			DockWorkspace.Instance.InitializeLayout(dockManager);
 
-			MessageBus<DockWorkspaceActiveTabPageChangedEventArgs>.Subscribers += DockWorkspace_ActiveTabPageChanged;
+			MenuService.Instance.Init(mainMenu, toolBar, InputBindings);
 
-			InitMainMenu();
-			InitWindowMenu();
-			InitToolbar();
 			InitFileLoaders();
 
 			this.Loaded += MainWindow_Loaded;
-		}
-
-		private void DockWorkspace_ActiveTabPageChanged(object sender, EventArgs e)
-		{
-			DockWorkspace dock = DockWorkspace.Instance;
-
-			var windowMenuItem = mainMenu.Items.OfType<MenuItem>().First(m => (string)m.Tag == nameof(Properties.Resources._Window));
-			foreach (MenuItem menuItem in windowMenuItem.Items.OfType<MenuItem>())
-			{
-				if (menuItem.IsCheckable && menuItem.Tag is TabPageModel)
-				{
-					menuItem.IsChecked = menuItem.Tag == dock.ActiveTabPage;
-				}
-			}
 		}
 
 		void SetWindowBounds(Rect bounds)
@@ -138,367 +107,6 @@ namespace ICSharpCode.ILSpy
 			this.Width = bounds.Width;
 			this.Height = bounds.Height;
 		}
-
-		#region Toolbar extensibility
-
-		void InitToolbar()
-		{
-			int navigationPos = 0;
-			int openPos = 1;
-			var toolbarCommands = App.ExportProvider.GetExports<ICommand, IToolbarCommandMetadata>("ToolbarCommand");
-			foreach (var commandGroup in toolbarCommands.OrderBy(c => c.Metadata.ToolbarOrder).GroupBy(c => Properties.Resources.ResourceManager.GetString(c.Metadata.ToolbarCategory)))
-			{
-				if (commandGroup.Key == Properties.Resources.ResourceManager.GetString("Navigation"))
-				{
-					foreach (var command in commandGroup)
-					{
-						toolBar.Items.Insert(navigationPos++, MakeToolbarItem(command));
-						openPos++;
-					}
-				}
-				else if (commandGroup.Key == Properties.Resources.ResourceManager.GetString("Open"))
-				{
-					foreach (var command in commandGroup)
-					{
-						toolBar.Items.Insert(openPos++, MakeToolbarItem(command));
-					}
-				}
-				else
-				{
-					toolBar.Items.Add(new Separator());
-					foreach (var command in commandGroup)
-					{
-						toolBar.Items.Add(MakeToolbarItem(command));
-					}
-				}
-			}
-
-		}
-
-		Button MakeToolbarItem(IExport<ICommand, IToolbarCommandMetadata> command)
-		{
-			return new Button {
-				Style = ThemeManager.Current.CreateToolBarButtonStyle(),
-				Command = CommandWrapper.Unwrap(command.Value),
-				ToolTip = Properties.Resources.ResourceManager.GetString(command.Metadata.ToolTip),
-				Tag = command.Metadata.Tag,
-				Content = new System.Windows.Controls.Image {
-					Width = 16,
-					Height = 16,
-					Source = Images.Load(command.Value, command.Metadata.ToolbarIcon)
-				}
-			};
-		}
-		#endregion
-
-		#region Main Menu extensibility
-
-		void InitMainMenu()
-		{
-			var mainMenuCommands = App.ExportProvider.GetExports<ICommand, IMainMenuCommandMetadata>("MainMenuCommand");
-			// Start by constructing the individual flat menus
-			var parentMenuItems = new Dictionary<string, MenuItem>();
-			var menuGroups = mainMenuCommands.OrderBy(c => c.Metadata.MenuOrder).GroupBy(c => c.Metadata.ParentMenuID);
-			foreach (var menu in menuGroups)
-			{
-				// Get or add the target menu item and add all items grouped by menu category
-				var parentMenuItem = GetOrAddParentMenuItem(menu.Key, menu.Key);
-				foreach (var category in menu.GroupBy(c => c.Metadata.MenuCategory))
-				{
-					if (parentMenuItem.Items.Count > 0)
-					{
-						parentMenuItem.Items.Add(new Separator { Tag = category.Key });
-					}
-					foreach (var entry in category)
-					{
-						if (menuGroups.Any(g => g.Key == entry.Metadata.MenuID))
-						{
-							var menuItem = GetOrAddParentMenuItem(entry.Metadata.MenuID, entry.Metadata.Header);
-							// replace potential dummy text with real name
-							menuItem.Header = GetResourceString(entry.Metadata.Header);
-							parentMenuItem.Items.Add(menuItem);
-						}
-						else
-						{
-							MenuItem menuItem = new MenuItem();
-							menuItem.Command = CommandWrapper.Unwrap(entry.Value);
-							menuItem.Tag = entry.Metadata.MenuID;
-							menuItem.Header = GetResourceString(entry.Metadata.Header);
-							if (!string.IsNullOrEmpty(entry.Metadata.MenuIcon))
-							{
-								menuItem.Icon = new Image {
-									Width = 16,
-									Height = 16,
-									Source = Images.Load(entry.Value, entry.Metadata.MenuIcon)
-								};
-							}
-
-							menuItem.IsEnabled = entry.Metadata.IsEnabled;
-							if (entry.Value is ToggleableCommand toggle)
-							{
-								menuItem.IsCheckable = true;
-								menuItem.SetBinding(MenuItem.IsCheckedProperty, new Binding("IsChecked") { Source = entry.Value, Mode = BindingMode.OneWay });
-							}
-
-							menuItem.InputGestureText = entry.Metadata.InputGestureText;
-							parentMenuItem.Items.Add(menuItem);
-						}
-					}
-				}
-			}
-
-			foreach (var (key, item) in parentMenuItems)
-			{
-				if (item.Parent == null)
-				{
-					mainMenu.Items.Add(item);
-				}
-			}
-
-			MenuItem GetOrAddParentMenuItem(string menuID, string resourceKey)
-			{
-				if (!parentMenuItems.TryGetValue(menuID, out var parentMenuItem))
-				{
-					var topLevelMenuItem = mainMenu.Items.OfType<MenuItem>().FirstOrDefault(m => (string)m.Tag == menuID);
-					if (topLevelMenuItem == null)
-					{
-						parentMenuItem = new MenuItem();
-						parentMenuItem.Header = GetResourceString(resourceKey);
-						parentMenuItem.Tag = menuID;
-						parentMenuItems.Add(menuID, parentMenuItem);
-					}
-					else
-					{
-						parentMenuItems.Add(menuID, topLevelMenuItem);
-						parentMenuItem = topLevelMenuItem;
-					}
-				}
-				return parentMenuItem;
-			}
-		}
-
-		internal static string GetResourceString(string key)
-		{
-			if (string.IsNullOrEmpty(key))
-			{
-				return null;
-			}
-			string value = Properties.Resources.ResourceManager.GetString(key);
-			if (!string.IsNullOrEmpty(value))
-			{
-				return value;
-			}
-			return key;
-		}
-		#endregion
-
-		#region Tool Pane extensibility
-
-		private void InitToolPanes()
-		{
-			var toolPanes = App.ExportProvider.GetExportedValues<ToolPaneModel>("ToolPane").OrderBy(item => item.Title);
-
-			DockWorkspace.Instance.ToolPanes.AddRange(toolPanes);
-		}
-
-		private void InitWindowMenu()
-		{
-			var windowMenuItem = mainMenu.Items.OfType<MenuItem>().First(m => (string)m.Tag == nameof(Properties.Resources._Window));
-			Separator separatorBeforeTools, separatorBeforeDocuments;
-			windowMenuItem.Items.Add(separatorBeforeTools = new Separator());
-			windowMenuItem.Items.Add(separatorBeforeDocuments = new Separator());
-
-			var dock = DockWorkspace.Instance;
-			dock.ToolPanes.CollectionChanged += ToolsChanged;
-			dock.TabPages.CollectionChanged += TabsChanged;
-
-			ToolsChanged(dock.ToolPanes, new NotifyCollectionChangedEventArgs(NotifyCollectionChangedAction.Reset));
-			TabsChanged(dock.TabPages, new NotifyCollectionChangedEventArgs(NotifyCollectionChangedAction.Reset));
-
-			void ToolsChanged(object sender, NotifyCollectionChangedEventArgs e)
-			{
-				int endIndex = windowMenuItem.Items.IndexOf(separatorBeforeDocuments);
-				int startIndex = windowMenuItem.Items.IndexOf(separatorBeforeTools) + 1;
-				int insertionIndex;
-				switch (e.Action)
-				{
-					case NotifyCollectionChangedAction.Add:
-						insertionIndex = Math.Min(endIndex, startIndex + e.NewStartingIndex);
-						foreach (ToolPaneModel pane in e.NewItems)
-						{
-							MenuItem menuItem = CreateMenuItem(pane);
-							windowMenuItem.Items.Insert(insertionIndex, menuItem);
-							insertionIndex++;
-						}
-						break;
-					case NotifyCollectionChangedAction.Remove:
-						foreach (ToolPaneModel pane in e.OldItems)
-						{
-							for (int i = endIndex - 1; i >= startIndex; i--)
-							{
-								MenuItem item = (MenuItem)windowMenuItem.Items[i];
-								if (pane == item.Tag)
-								{
-									windowMenuItem.Items.RemoveAt(i);
-									item.Tag = null;
-									endIndex--;
-									break;
-								}
-							}
-						}
-						break;
-					case NotifyCollectionChangedAction.Replace:
-						break;
-					case NotifyCollectionChangedAction.Move:
-						break;
-					case NotifyCollectionChangedAction.Reset:
-						for (int i = endIndex - 1; i >= startIndex; i--)
-						{
-							MenuItem item = (MenuItem)windowMenuItem.Items[0];
-							item.Tag = null;
-							windowMenuItem.Items.RemoveAt(i);
-							endIndex--;
-						}
-						insertionIndex = endIndex;
-						foreach (ToolPaneModel pane in dock.ToolPanes)
-						{
-							MenuItem menuItem = CreateMenuItem(pane);
-							windowMenuItem.Items.Insert(insertionIndex, menuItem);
-							insertionIndex++;
-						}
-						break;
-				}
-
-				MenuItem CreateMenuItem(ToolPaneModel pane)
-				{
-					MenuItem menuItem = new MenuItem();
-					menuItem.Command = pane.AssociatedCommand ?? new ToolPaneCommand(pane.ContentId);
-					menuItem.Header = pane.Title;
-					menuItem.Tag = pane;
-					var shortcutKey = pane.ShortcutKey;
-					if (shortcutKey != null)
-					{
-						InputBindings.Add(new InputBinding(menuItem.Command, shortcutKey));
-						menuItem.InputGestureText = shortcutKey.GetDisplayStringForCulture(CultureInfo.CurrentUICulture);
-					}
-					if (!string.IsNullOrEmpty(pane.Icon))
-					{
-						menuItem.Icon = new Image {
-							Width = 16,
-							Height = 16,
-							Source = Images.Load(pane, pane.Icon)
-						};
-					}
-
-					return menuItem;
-				}
-			}
-
-			void TabsChanged(object sender, NotifyCollectionChangedEventArgs e)
-			{
-				int endIndex = windowMenuItem.Items.Count;
-				int startIndex = windowMenuItem.Items.IndexOf(separatorBeforeDocuments) + 1;
-				int insertionIndex;
-				switch (e.Action)
-				{
-					case NotifyCollectionChangedAction.Add:
-						insertionIndex = Math.Min(endIndex, startIndex + e.NewStartingIndex);
-						foreach (TabPageModel pane in e.NewItems)
-						{
-							MenuItem menuItem = CreateMenuItem(pane);
-							pane.PropertyChanged += TabPageChanged;
-							windowMenuItem.Items.Insert(insertionIndex, menuItem);
-							insertionIndex++;
-						}
-						break;
-					case NotifyCollectionChangedAction.Remove:
-						foreach (TabPageModel pane in e.OldItems)
-						{
-							for (int i = endIndex - 1; i >= startIndex; i--)
-							{
-								MenuItem item = (MenuItem)windowMenuItem.Items[i];
-								if (pane == item.Tag)
-								{
-									windowMenuItem.Items.RemoveAt(i);
-									pane.PropertyChanged -= TabPageChanged;
-									item.Tag = null;
-									endIndex--;
-									break;
-								}
-							}
-						}
-						break;
-					case NotifyCollectionChangedAction.Replace:
-						break;
-					case NotifyCollectionChangedAction.Move:
-						break;
-					case NotifyCollectionChangedAction.Reset:
-						for (int i = endIndex - 1; i >= startIndex; i--)
-						{
-							MenuItem item = (MenuItem)windowMenuItem.Items[i];
-							windowMenuItem.Items.RemoveAt(i);
-							((TabPageModel)item.Tag).PropertyChanged -= TabPageChanged;
-							endIndex--;
-						}
-						insertionIndex = endIndex;
-						foreach (TabPageModel pane in dock.TabPages)
-						{
-							MenuItem menuItem = CreateMenuItem(pane);
-							pane.PropertyChanged += TabPageChanged;
-							windowMenuItem.Items.Insert(insertionIndex, menuItem);
-							insertionIndex++;
-						}
-						break;
-				}
-
-				MenuItem CreateMenuItem(TabPageModel pane)
-				{
-					MenuItem menuItem = new MenuItem();
-					menuItem.Command = new TabPageCommand(pane);
-					menuItem.Header = pane.Title.Length > 20 ? pane.Title.Substring(20) + "..." : pane.Title;
-					menuItem.Tag = pane;
-					menuItem.IsCheckable = true;
-
-					return menuItem;
-				}
-			}
-
-			static void TabPageChanged(object sender, PropertyChangedEventArgs e)
-			{
-				var windowMenuItem = Instance.mainMenu.Items.OfType<MenuItem>().First(m => (string)m.Tag == nameof(Properties.Resources._Window));
-				foreach (MenuItem menuItem in windowMenuItem.Items.OfType<MenuItem>())
-				{
-					if (menuItem.IsCheckable && menuItem.Tag == sender)
-					{
-						string title = ((TabPageModel)sender).Title;
-						menuItem.Header = title.Length > 20 ? title.Substring(0, 20) + "..." : title;
-					}
-				}
-			}
-		}
-
-		public void ShowInTopPane(string title, object content)
-		{
-			var model = DockWorkspace.Instance.ToolPanes.OfType<LegacyToolPaneModel>().FirstOrDefault(p => p.Content == content);
-			if (model == null)
-			{
-				model = new LegacyToolPaneModel(title, content, LegacyToolPaneLocation.Top);
-				DockWorkspace.Instance.ToolPanes.Add(model);
-			}
-			model.Show();
-		}
-
-		public void ShowInBottomPane(string title, object content)
-		{
-			var model = DockWorkspace.Instance.ToolPanes.OfType<LegacyToolPaneModel>().FirstOrDefault(p => p.Content == content);
-			if (model == null)
-			{
-				model = new LegacyToolPaneModel(title, content, LegacyToolPaneLocation.Bottom);
-				DockWorkspace.Instance.ToolPanes.Add(model);
-			}
-			model.Show();
-		}
-		#endregion
 
 		#region File Loader extensibility
 
@@ -514,20 +122,22 @@ namespace ICSharpCode.ILSpy
 		#endregion
 
 		#region Message Hook
+
 		protected override void OnSourceInitialized(EventArgs e)
 		{
 			base.OnSourceInitialized(e);
-			PresentationSource source = PresentationSource.FromVisual(this);
+
+			var source = PresentationSource.FromVisual(this);
 
 			var sessionSettings = SettingsService.Instance.SessionSettings;
 
 			// Validate and Set Window Bounds
-			Rect bounds = Rect.Transform(sessionSettings.WindowBounds, source.CompositionTarget.TransformToDevice);
-			var boundsRect = new System.Drawing.Rectangle((int)bounds.Left, (int)bounds.Top, (int)bounds.Width, (int)bounds.Height);
+			Rect bounds = Rect.Transform(sessionSettings.WindowBounds, source?.CompositionTarget?.TransformToDevice ?? Matrix.Identity);
+			var boundsRect = new Rectangle((int)bounds.Left, (int)bounds.Top, (int)bounds.Width, (int)bounds.Height);
 			bool boundsOK = false;
-			foreach (var screen in System.Windows.Forms.Screen.AllScreens)
+			foreach (var screen in Screen.AllScreens)
 			{
-				var intersection = System.Drawing.Rectangle.Intersect(boundsRect, screen.WorkingArea);
+				var intersection = Rectangle.Intersect(boundsRect, screen.WorkingArea);
 				if (intersection.Width > 10 && intersection.Height > 10)
 					boundsOK = true;
 			}
@@ -852,24 +462,6 @@ namespace ICSharpCode.ILSpy
 				return null;
 
 			return loadedAssy.FileName;
-		}
-
-		public void SetStatus(string status, Brush foreground)
-		{
-			if (this.statusBar.Visibility == Visibility.Collapsed)
-				this.statusBar.Visibility = Visibility.Visible;
-			this.statusLabel.Foreground = foreground;
-			this.statusLabel.Text = status;
-		}
-
-		public ItemCollection GetMainMenuItems()
-		{
-			return mainMenu.Items;
-		}
-
-		public ItemCollection GetToolBarItems()
-		{
-			return toolBar.Items;
 		}
 	}
 }
